refactor(prover): delete EpochProver / ServerEpochProver and stop exporting ProvingOrchestrator
Now that EpochProvingJob talks to CheckpointSubTreeOrchestrator and
TopTreeOrchestrator directly, the wrapper layer has no production users:

- Delete the EpochProver interface (stdlib/src/interfaces/epoch-prover.ts)
  and its re-export from interfaces/server.ts.
- Delete ServerEpochProver, the adapter that translated EpochProver calls
  onto ProvingOrchestrator + a broker facade.
- Drop createEpochProver from EpochProverManager and from ProverClient.
  ProverClient now exposes only the split factories.
- Drop ProvingOrchestrator from prover-client/orchestrator's package
  exports, and remove its `implements EpochProver` clause. The class file
  stays as the base for CheckpointSubTreeOrchestrator (which extends it)
  and as the single-class end-to-end driver used by orchestrator_*.test.ts;
  it is no longer reachable from outside the package.
- Switch test_context.ts to import ProvingOrchestrator via its relative
  module path (the orchestrator-internal test driver TestProvingOrchestrator
  still extends it).

All 301 prover-client tests, 87 prover-node tests, and 801 stdlib tests
still pass.
