+++ yarn-project/prover-client/src/orchestrator/index.ts
@@ -1,4 +1,3 @@
-export { ProvingOrchestrator } from './orchestrator.js';
 export { CheckpointSubTreeOrchestrator, type SubTreeResult } from './checkpoint-sub-tree-orchestrator.js';
 export {
   TopTreeOrchestrator,
