+++ yarn-project/stdlib/src/interfaces/prover-client.ts
@@ -5,7 +5,6 @@ import { z } from 'zod';
 
 import { schemas, zodFor } from '../schemas/index.js';
 import type { TxHash } from '../tx/tx_hash.js';
-import type { EpochProver } from './epoch-prover.js';
 import type { ProvingJobConsumer } from './prover-broker.js';
 
 export type ActualProverConfig = {
@@ -131,8 +130,6 @@ function parseProverId(str: string) {
  * `prover-client` itself. Consumers (e.g. `EpochProvingJob`) import the concrete class.
  */
 export interface EpochProverManager {
-  createEpochProver(): EpochProver;
-
   start(): Promise<void>;
 
   stop(): Promise<void>;
